requirements.txt
altair==5.3.0
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
facenet-pytorch==2.5.3
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
googledriver==0.1.7
grad-cam==1.5.0
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.34.0
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
ttach==0.0.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uv==0.1.44
watchdog==4.0.0
wheel==0.43.0
opencv_python_headless==4.9.0.80
